Refactored the code.

+++ include/benchmarks/gemm/gemm_blaze.h
@@ -3,16 +3,16 @@
 
 #include "benchmark/benchmark.h"
 #include "blaze/Math.h"
-#include "gemm_header.h"
+#include "shared/common.hpp"
 #include "random.hpp"
 
 static void gemm_blaze(benchmark::State& state) {
-    blaze::StaticMatrix<double, SIZE, SIZE, blaze::rowMajor> m1, m2, res;
+    blaze::StaticMatrix<float, SIZE, SIZE, blaze::rowMajor> m1, m2, res;
 
     m1 =
-        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomDouble(-1.0, 1.0); });
+        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomFloat(-1.0, 1.0); });
     m2 =
-        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomDouble(-1.0, 1.0); });
+        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomFloat(-1.0, 1.0); });
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(res);
