+++ include/benchmarks/gemm/gemm_eigen.h
@@ -3,13 +3,13 @@
 
 #include "Eigen/Dense"
 #include "benchmark/benchmark.h"
-#include "gemm_header.h"
+#include "shared/common.hpp"
 
 static void gemm_eigen(benchmark::State& state) {
-    Eigen::Matrix<double, SIZE, SIZE> m1, m2, res;
+    Eigen::Matrix<float, SIZE, SIZE> m1, m2, res;
 
-    m1 = Eigen::Matrix<double, SIZE, SIZE>::Random();
-    m2 = Eigen::Matrix<double, SIZE, SIZE>::Random();
+    m1 = Eigen::Matrix<float, SIZE, SIZE>::Random();
+    m2 = Eigen::Matrix<float, SIZE, SIZE>::Random();
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(res);
